fold the pamphlet stuff into GuiBookLanding
